requirements.txt
certifi==2021.10.8
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
imutils==0.5.4
itsdangerous==2.1.0
Jinja2==3.0.3
numpy==1.18.5
opencv-python==4.5.5.62
pandas==1.4.1
pdf2image==1.16.0
pymongo==4.0.1
pytesseract==0.3.9
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
urllib3==1.26.8